Add StompProtocol
